updated review comments
